refactor: extract chunk_text/chunk_json to shared stream_utils module
